Move files around and create Rakefile
Since Weechat can be picky about requires stuff, we just have a rake
task that cats everything into one big script to be loaded into weechat.
Hacky, but it works.

This commit introduces the /moon command, which when given the date will
return the approximate illumination of the moon for that date.